Derive list of policy files in Makefile
This is better than a hard coded list. Also it picks up the new
redhat config, which was being skipped from the CI.

Also, remove duplicate exclude for step_image_registries in the
redhat config. The exclude is in the template so it's added to all
the policy files, that's why we don't need to specify it in the
source list.